Move most of //sky/engine/core/painting to //flutter/lib/ui/painting
These parts don't depend on engine anymore. Also reformat them to match Google
C++ style.
