+++ flutter/lib/ui/BUILD.gn
@@ -1,0 +1,36 @@
+# Copyright 2016 The Chromium Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+source_set("ui") {
+  sources = [
+    "painting/color_filter.cc",
+    "painting/color_filter.h",
+    "painting/gradient.cc",
+    "painting/gradient.h",
+    "painting/image.cc",
+    "painting/image_filter.cc",
+    "painting/image_filter.h",
+    "painting/image.h",
+    "painting/image_shader.cc",
+    "painting/image_shader.h",
+    "painting/mask_filter.cc",
+    "painting/mask_filter.h",
+    "painting/matrix.cc",
+    "painting/matrix.h",
+    "painting/path.cc",
+    "painting/path.h",
+    "painting/picture.cc",
+    "painting/picture.h",
+    "painting/rrect.cc",
+    "painting/rrect.h",
+    "painting/shader.cc",
+    "painting/shader.h",
+  ]
+
+  deps = [
+    "//base",
+    "//flutter/tonic",
+    "//skia",
+  ]
+}
