+++ flutter/lib/ui/painting/image.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef SKY_ENGINE_CORE_PAINTING_CANVASIMAGE_H_
-#define SKY_ENGINE_CORE_PAINTING_CANVASIMAGE_H_
+#ifndef FLUTTER_LIB_UI_PAINTING_IMAGE_H_
+#define FLUTTER_LIB_UI_PAINTING_IMAGE_H_
 
 #include "base/memory/ref_counted.h"
 #include "flutter/tonic/dart_wrappable.h"
@@ -17,14 +17,14 @@ class CanvasImage final : public base::RefCountedThreadSafe<CanvasImage>,
   DEFINE_WRAPPERTYPEINFO();
  public:
   ~CanvasImage() override;
-  static scoped_refptr<CanvasImage> create() { return new CanvasImage(); }
+  static scoped_refptr<CanvasImage> Create() { return new CanvasImage(); }
 
-  int width();
-  int height();
+  int width() { return image_->width(); }
+  int height() { return image_->height(); }
   void dispose();
 
-  sk_sp<SkImage> image() const { return image_; }
-  void setImage(sk_sp<SkImage> image) { image_ = image; }
+  const sk_sp<SkImage>& image() const { return image_; }
+  void set_image(sk_sp<SkImage> image) { image_ = std::move(image); }
 
   static void RegisterNatives(DartLibraryNatives* natives);
 
@@ -36,4 +36,4 @@ class CanvasImage final : public base::RefCountedThreadSafe<CanvasImage>,
 
 }  // namespace blink
 
-#endif  // SKY_ENGINE_CORE_PAINTING_CANVASIMAGE_H_
+#endif  // FLUTTER_LIB_UI_PAINTING_IMAGE_H_
