+++ flutter/lib/ui/painting/mask_filter.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/engine/core/painting/MaskFilter.h"
+#include "flutter/lib/ui/painting/mask_filter.h"
 
 #include "flutter/tonic/dart_args.h"
 #include "flutter/tonic/dart_binding_macros.h"
@@ -13,7 +13,7 @@
 namespace blink {
 
 static void MaskFilter_constructor(Dart_NativeArguments args) {
-  DartCallConstructor(&MaskFilter::create, args);
+  DartCallConstructor(&MaskFilter::Create, args);
 }
 
 IMPLEMENT_WRAPPERTYPEINFO(ui, MaskFilter);
@@ -24,14 +24,14 @@ void MaskFilter::RegisterNatives(DartLibraryNatives* natives) {
   });
 }
 
-scoped_refptr<MaskFilter> MaskFilter::create(
-      unsigned style, double sigma, unsigned flags) {
+scoped_refptr<MaskFilter> MaskFilter::Create(unsigned style, double sigma,
+                                             unsigned flags) {
   return new MaskFilter(SkBlurMaskFilter::Make(
       static_cast<SkBlurStyle>(style), sigma, flags));
 }
 
 MaskFilter::MaskFilter(sk_sp<SkMaskFilter> filter)
-    : filter_(filter) {
+  : filter_(std::move(filter)) {
 }
 
 MaskFilter::~MaskFilter() {
