+++ flutter/lib/ui/painting/matrix.h
@@ -2,17 +2,17 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef SKY_ENGINE_CORE_PAINTING_MATRIX_H_
-#define SKY_ENGINE_CORE_PAINTING_MATRIX_H_
+#ifndef FLUTTER_LIB_UI_PAINTING_MATRIX_H_
+#define FLUTTER_LIB_UI_PAINTING_MATRIX_H_
 
 #include "flutter/tonic/float64_list.h"
 #include "third_party/skia/include/core/SkMatrix.h"
 
 namespace blink {
 
-SkMatrix toSkMatrix(const Float64List& matrix4);
-Float64List toMatrix4(const SkMatrix& sk_matrix);
+SkMatrix ToSkMatrix(const Float64List& matrix4);
+Float64List ToMatrix4(const SkMatrix& sk_matrix);
 
 }  // namespace blink
 
-#endif  // SKY_ENGINE_CORE_PAINTING_MATRIX_H_
+#endif  // FLUTTER_LIB_UI_PAINTING_MATRIX_H_
