+++ flutter/lib/ui/painting/rrect.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/engine/core/painting/RRect.h"
+#include "flutter/lib/ui/painting/rrect.h"
 
 #include "flutter/tonic/dart_error.h"
 #include "flutter/tonic/float32_list.h"
