+++ sky/engine/bindings/dart_ui.cc
@@ -4,6 +4,14 @@
 
 #include "sky/engine/bindings/dart_ui.h"
 
+#include "flutter/lib/ui/painting/color_filter.h"
+#include "flutter/lib/ui/painting/gradient.h"
+#include "flutter/lib/ui/painting/image.h"
+#include "flutter/lib/ui/painting/image_filter.h"
+#include "flutter/lib/ui/painting/image_shader.h"
+#include "flutter/lib/ui/painting/mask_filter.h"
+#include "flutter/lib/ui/painting/path.h"
+#include "flutter/lib/ui/painting/picture.h"
 #include "flutter/tonic/dart_converter.h"
 #include "flutter/tonic/dart_error.h"
 #include "sky/engine/bindings/dart_runtime_hooks.h"
@@ -11,14 +19,6 @@
 #include "sky/engine/core/compositing/Scene.h"
 #include "sky/engine/core/compositing/SceneBuilder.h"
 #include "sky/engine/core/painting/Canvas.h"
-#include "sky/engine/core/painting/CanvasGradient.h"
-#include "sky/engine/core/painting/CanvasImage.h"
-#include "sky/engine/core/painting/CanvasPath.h"
-#include "sky/engine/core/painting/ColorFilter.h"
-#include "sky/engine/core/painting/ImageFilter.h"
-#include "sky/engine/core/painting/ImageShader.h"
-#include "sky/engine/core/painting/MaskFilter.h"
-#include "sky/engine/core/painting/Picture.h"
 #include "sky/engine/core/painting/PictureRecorder.h"
 #include "sky/engine/core/painting/painting.h"
 #include "sky/engine/core/text/Paragraph.h"
