+++ sky/engine/core/core.gni
@@ -20,34 +20,12 @@ sky_core_files = [
   "editing/PositionWithAffinity.h",
   "painting/Canvas.cpp",
   "painting/Canvas.h",
-  "painting/CanvasGradient.cpp",
-  "painting/CanvasGradient.h",
-  "painting/CanvasImage.cpp",
-  "painting/CanvasImage.h",
-  "painting/CanvasPath.cpp",
-  "painting/CanvasPath.h",
-  "painting/ColorFilter.cpp",
-  "painting/ColorFilter.h",
-  "painting/ImageFilter.cpp",
-  "painting/ImageFilter.h",
-  "painting/ImageShader.cpp",
-  "painting/ImageShader.h",
-  "painting/MaskFilter.cpp",
-  "painting/MaskFilter.h",
-  "painting/Matrix.cpp",
-  "painting/Matrix.h",
-  "painting/painting.cc",
-  "painting/painting.h",
   "painting/Paint.cpp",
   "painting/Paint.h",
-  "painting/Picture.cpp",
-  "painting/Picture.h",
+  "painting/painting.cc",
+  "painting/painting.h",
   "painting/PictureRecorder.cpp",
   "painting/PictureRecorder.h",
-  "painting/RRect.cpp",
-  "painting/RRect.h",
-  "painting/Shader.cpp",
-  "painting/Shader.h",
   "rendering/BidiRun.h",
   "rendering/BidiRunForLine.cpp",
   "rendering/BidiRunForLine.h",
@@ -219,13 +197,7 @@ core_dart_files = get_path_info([
                                   "dart/lerp.dart",
                                   "dart/mojo_services.dart",
                                   "dart/natives.dart",
-                                  "dart/painting.dart",
                                   "dart/text.dart",
                                   "dart/window.dart",
-                                  "painting/Color.dart",
-                                  "painting/FilterQuality.dart",
-                                  "painting/Paint.dart",
-                                  "painting/PaintingStyle.dart",
-                                  "painting/TransferMode.dart",
                                 ],
                                 "abspath")
