--- sky/engine/core/painting/FilterQuality.dart
@@ -1,34 +1,0 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-part of dart_ui;
-
-// List of predefined filter quality modes. This list comes from Skia's
-// SkFilterQuality.h and the values (order) should be kept in sync.
-
-/// Quality levels for image filters.
-///
-/// See [Paint.filterQuality].
-enum FilterQuality {
-  /// Fastest possible filtering, albeit also the lowest quality.
-  ///
-  /// Typically this implies nearest-neighbour filtering.
-  none,
-
-  /// Better quality than [none], faster than [medium].
-  ///
-  /// Typically this implies bilinear interpolation.
-  low,
-
-  /// Better quality than [low], faster than [high].
-  ///
-  /// Typically this implies a combination of bilinear interpolation and
-  /// pyramidal parametric prefiltering (mipmaps).
-  medium,
-
-  /// Best possible quality filtering, albeit also the slowest.
-  ///
-  /// Typically this implies bicubic interpolation or better.
-  high,
-}
